chromium/third_party/blink/web_tests/virtual/out-of-blink-cspee/external/wpt/content-security-policy/embedded-enforcement/subsumption_algorithm-strict_dynamic-expected.txt

This is a testharness.js-based test.
PASS 'strict-dynamic' is ineffective for `style-src`.
PASS 'strict-dynamic' is ineffective for `img-src`.
PASS 'strict-dynamic' is ineffective for `frame-src`.
PASS 'strict-dynamic' is ineffective for `child-src`.
FAIL 'strict-dynamic' is effective only for `script-src`. assert_equals: expected (undefined) undefined but got (boolean) true
FAIL 'strict-dynamic' is properly handled for finding effective policy. assert_equals: expected (undefined) undefined but got (boolean) true
PASS 'strict-dynamic' makes host source expressions ineffective.
PASS 'strict-dynamic' makes scheme source expressions ineffective.
PASS 'strict-dynamic' makes 'self' ineffective.
PASS 'strict-dynamic' makes 'unsafe-inline' ineffective.
FAIL 'strict-dynamic' has to be allowed by required csp if it is present in returned csp. assert_unreached: No message should be sent from the frame. Reached unreachable code
Harness: the test ran to completion.
